requirements.txt
alembic==1.4.3
aniso8601==8.1.0
Authlib==0.15.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
pycparser==2.20
pycryptodome==3.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.4
urllib3==1.26.2
Werkzeug==0.15.4
wrapt==1.11.1
